Add files via upload
There are five classes with different number of students in them. Five arrays containing marks of students of  each class. Write a function to recieve an adderess of an array of pointers to access marks of all students. The job of function is to find the highest marks among all the classes.